backend/requirements.txt
flask
flask-cors
torch
torchaudio
openai-whisper
numpy
